Move Wizard and MigrateModal into plans page components directory
