+++ src/app/home/pages/PlansPage/components/Wizard/ResultsStep.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import spacing from '@patternfly/react-styles/css/utilities/Spacing/spacing';
 import { Bullseye, Title, Button, Flex, FlexModifiers, FlexItem } from '@patternfly/react-core';
 import ConditionsGrid from './ConditionsGrid';
-import { ICurrentPlanStatus, CurrentPlanState } from '../../duck/reducers';
+import { ICurrentPlanStatus, CurrentPlanState } from '../../../../../plan/duck/reducers';
 import { Spinner } from '@patternfly/react-core';
 import {
   IconSize,
